python_tutor/pyproject.toml
[project]
name = "neuroscience-python-tutor"
version = "0.1.0"
description = "A graduate-level Python RAG tutor for research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.3.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.1.0",
]

[tool.uv]
# This ensures a reproducible environment for all students
managed = true
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"